Separate out processor module in cirrus-executor
Add TODO: shuffle_extrinsics
