+++ cumulus/client/cirrus-service/src/lib.rs
@@ -103,7 +103,7 @@ where
 		+ BlockchainEvents<Block>
 		+ ProvideRuntimeApi<Block>
 		+ 'static,
-    Client::Api: cirrus_primitives::SecondaryApi<Block, cirrus_primitives::AccountId>,
+	Client::Api: cirrus_primitives::SecondaryApi<Block, cirrus_primitives::AccountId>,
 	RClient: RelaychainClient + Clone + Send + Sync + 'static,
 	for<'b> &'b Client: BlockImport<Block>,
 	Spawner: SpawnNamed + Clone + Send + Sync + 'static,
